+++ src/app/api/chat/[provider]/agentRuntime.ts
@@ -1,8 +1,15 @@
 import { getServerConfig } from '@/config/server';
 import { JWTPayload } from '@/const/auth';
+import { INBOX_SESSION_ID } from '@/const/session';
 import {
-  ChatCompetitionOptions,
+  LOBE_CHAT_OBSERVATION_ID,
+  LOBE_CHAT_TRACE_ID,
+  TracePayload,
+  TraceTagMap,
+} from '@/const/trace';
+import {
   ChatStreamPayload,
+  LobeAnthropicAI,
   LobeAzureOpenAI,
   LobeBedrockAI,
   LobeGoogleAI,
@@ -14,6 +21,7 @@ import {
   LobeZhipuAI,
   ModelProvider,
 } from '@/libs/agent-runtime';
+import { TraceClient } from '@/libs/traces';
 
 import apiKeyManager from '../apiKeyManager';
 
@@ -23,6 +31,12 @@ interface AzureOpenAIParams {
   useAzure?: boolean;
 }
 
+export interface AgentChatOptions {
+  enableTrace?: boolean;
+  provider: string;
+  trace?: TracePayload;
+}
+
 class AgentRuntime {
   private _runtime: LobeRuntimeAI;
 
@@ -30,8 +44,67 @@ class AgentRuntime {
     this._runtime = runtime;
   }
 
-  async chat(payload: ChatStreamPayload, options?: ChatCompetitionOptions) {
-    return this._runtime.chat(payload, options);
+  async chat(
+    payload: ChatStreamPayload,
+    { trace: tracePayload, provider, enableTrace }: AgentChatOptions,
+  ) {
+    const { messages, model, tools, ...parameters } = payload;
+
+    // if not enabled trace then just call the runtime
+    if (!enableTrace) return this._runtime.chat(payload);
+
+    // create a trace to monitor the completion
+    const traceClient = new TraceClient();
+    const trace = traceClient.createTrace({
+      id: tracePayload?.traceId,
+      input: messages,
+      metadata: { provider },
+      name: tracePayload?.traceName,
+      sessionId: `${tracePayload?.sessionId || INBOX_SESSION_ID}@${tracePayload?.topicId || 'start'}`,
+      tags: tracePayload?.tags,
+      userId: tracePayload?.userId,
+    });
+
+    const generation = trace?.generation({
+      input: messages,
+      metadata: { provider },
+      model,
+      modelParameters: parameters as any,
+      name: `Chat Completion (${provider})`,
+      startTime: new Date(),
+    });
+
+    return this._runtime.chat(payload, {
+      callback: {
+        experimental_onToolCall: async () => {
+          trace?.update({
+            tags: [...(tracePayload?.tags || []), TraceTagMap.ToolsCall],
+          });
+        },
+
+        onCompletion: async (completion) => {
+          generation?.update({
+            endTime: new Date(),
+            metadata: { provider, tools },
+            output: completion,
+          });
+
+          trace?.update({ output: completion });
+        },
+
+        onFinal: async () => {
+          await traceClient.shutdownAsync();
+        },
+
+        onStart: () => {
+          generation?.update({ completionStartTime: new Date() });
+        },
+      },
+      headers: {
+        [LOBE_CHAT_OBSERVATION_ID]: generation?.id,
+        [LOBE_CHAT_TRACE_ID]: trace?.id,
+      },
+    });
   }
 
   static async initializeWithUserPayload(
@@ -81,6 +154,11 @@ class AgentRuntime {
 
       case ModelProvider.Perplexity: {
         runtimeModel = this.initPerplexity(payload);
+        break;
+      }
+
+      case ModelProvider.Anthropic: {
+        runtimeModel = this.initAnthropic(payload);
         break;
       }
     }
@@ -169,6 +247,13 @@ class AgentRuntime {
     const apiKey = apiKeyManager.pick(payload?.apiKey || PERPLEXITY_API_KEY);
 
     return new LobePerplexityAI({ apiKey });
+  }
+
+  private static initAnthropic(payload: JWTPayload) {
+    const { ANTHROPIC_API_KEY } = getServerConfig();
+    const apiKey = apiKeyManager.pick(payload?.apiKey || ANTHROPIC_API_KEY);
+
+    return new LobeAnthropicAI({ apiKey });
   }
 }