+++ src/services/file.ts
@@ -34,8 +34,7 @@ class FileService {
       }))();
 
     // 压缩图片
-    const fileType = 'image/webp';
-    const base64String = compressImage({ img, type: fileType });
+    const base64String = compressImage({ img, type: file.fileType });
     const binaryString = atob(base64String.split('base64,')[1]);
     const uint8Array = Uint8Array.from(binaryString, (char) => char.charCodeAt(0));
     file.data = uint8Array.buffer;